// Qanat.Web/src/app/shared/generated/model/create-water-account-from-suggestion-dto.ts
/**
 * Qanat.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class CreateWaterAccountFromSuggestionDto { 
    ParcelIDList: Array<number>;
    ReportingPeriodID: number;
    WaterAccountName: string;
    ContactName?: string | null;
    ContactAddress?: string | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface CreateWaterAccountFromSuggestionDtoForm { 
    ParcelIDList: FormControl<Array<number>>;
    ReportingPeriodID: FormControl<number>;
    WaterAccountName: FormControl<string>;
    ContactName?: FormControl<string>;
    ContactAddress?: FormControl<string>;
}

export class CreateWaterAccountFromSuggestionDtoFormControls { 
    public static ParcelIDList = (value: FormControlState<Array<number>> | Array<number> = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<Array<number>>(
        value,
        formControlOptions ?? 
        {
            nonNullable: true,
            validators: 
            [
                Validators.required,
            ],
        }
    );
    public static ReportingPeriodID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: true,
            validators: 
            [
                Validators.required,
            ],
        }
    );
    public static WaterAccountName = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: true,
            validators: 
            [
                Validators.required,
                Validators.minLength(1),
            ],
        }
    );
    public static ContactName = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static ContactAddress = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}