// Qanat.Web/src/app/shared/generated/model/water-account-suggestion-dto.ts
export class WaterAccountSuggestionDto { 
    WaterAccountName?: string | null;
    ParcelNumbers?: string | null;
    ParcelIDList?: string | null;
    WellIDList?: string | null;
    WellIDs?: Array<WellLinkDisplayDto> | null;
    ContactName?: string | null;
    ContactAddress?: string | null;
    ParcelArea?: number;
    Zones?: string | null;
    Parcels?: Array<ParcelLinkDisplayDto> | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}